Tentatively made the `flag` variable an i32
This has been a problem that inserting release flags > 3 cause `flag` to
overflow. It should now be written as a 32 bit integer, and the 2020.0.C
to 2020.1.A conversion should also convert it if needed.

+++ write_tccon_netcdf/write_netcdf.py
@@ -2375,7 +2375,7 @@ def main_inner(args, logger, show_progress, HEAD_commit, code_dir, GGGPATH):
 
         # flags
         logging.info('\t- Flags')
-        nc_data.createVariable('flag',np.int16,('time',))
+        nc_data.createVariable('flag',np.int32,('time',))
         att_dict = {
             "description": 'data quality flag, 0 = good',
             "standard_name": 'quality_flag',
